More functions. Second demo.
git-svn-id: https://svn.code.sf.net/p/curses-sharp/code/trunk@10 52a65036-936b-447d-8c72-270fba66c20f

+++ CursesSharp/Defs.cs
@@ -9,7 +9,7 @@ namespace Curses
 			return KEY_F0 + n;
 		}
 
-        public static uint COLOR_PAIR(short n)
+        public static uint COLOR_PAIR(int n)
         {
             return NativeMethods.wrap_COLOR_PAIR(n);
         }
